make iterator into template and move common RDD code into base class and remove copied code
